EPMLABSBRN-669 renamed LopotkoProcessorTest for the increasing of coverage
